FALL-DETECTION/requirements.txt
ultralytics
cvzone
opencv-python
requests
python-dotenv
cloudinary
mediapipe
numpy